backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
anthropic==0.43.0
asyncpg==0.29.0
python-dotenv==1.0.0
apscheduler==3.10.4
pydantic==2.5.3
